Encounter Ideas 1.5
ALL NEED TESTING. Just went through a fixed a lot of errors that I noticed. I also was struggling to reformat because my online compiler was indenting lines improperly. So if the formatting is HELLA JANK, blame the compiler switching to GitHub not me! 

Some of these encounters have enemy with special logic. I've left that logic in the Enemy Ideas! :)